Separate Country and CountryGroup classes into separate files
