+++ support-frontend/assets/components/personalDetails/stateSelect.tsx
@@ -1,5 +1,6 @@
 import { Option, Select } from '@guardian/source-react-components';
 import type { ContributionType } from 'helpers/contributions';
+import { CountryGroup } from 'helpers/internationalisation';
 import type { IsoCountry } from 'helpers/internationalisation/country';
 import {
 	auStates,
@@ -7,7 +8,6 @@ import {
 	usStates,
 } from 'helpers/internationalisation/country';
 import type { CountryGroupId } from 'helpers/internationalisation/countryGroup';
-import { CountryGroup } from 'helpers/internationalisation/helpers';
 import { shouldCollectStateForContributions } from 'helpers/internationalisation/shouldCollectStateForContribs';
 
 type StateSelectProps = {
