+++ support-frontend/assets/helpers/internationalisation/index.ts
@@ -1,0 +1,4 @@
+import Country from './classes/country';
+import CountryGroup from './classes/countryGroup';
+
+export { Country, CountryGroup };
