+++ support-frontend/assets/helpers/redux/checkout/address/state.ts
@@ -1,7 +1,7 @@
 import { z } from 'zod';
 import type { PostcodeFinderResult } from 'components/subscriptionCheckouts/address/postcodeLookup';
+import { Country } from 'helpers/internationalisation';
 import { isoCountries } from 'helpers/internationalisation/country';
-import { Country } from 'helpers/internationalisation/helpers';
 import type { SliceErrors } from 'helpers/redux/utils/validation/errors';
 import type { FormError } from 'helpers/subscriptionsForms/validation';
 import { isPostCodeValid } from './validationFunctions';
