+++ support-frontend/assets/helpers/redux/commonState/reducer.ts
@@ -1,9 +1,9 @@
 import type { PayloadAction } from '@reduxjs/toolkit';
 import { createSlice } from '@reduxjs/toolkit';
 import type { ContributionTypes } from 'helpers/contributions';
+import { CountryGroup } from 'helpers/internationalisation';
 import type { IsoCountry } from 'helpers/internationalisation/country';
 import { fromCountryGroupId } from 'helpers/internationalisation/currency';
-import { CountryGroup } from 'helpers/internationalisation/helpers';
 import type { Participations } from '../../abTests/abtest';
 import type { CommonStateSetupData, Internationalisation } from './state';
 import { initialCommonState } from './state';
