+++ support-frontend/assets/helpers/redux/commonState/state.ts
@@ -4,11 +4,11 @@ import type { SelectedAmountsVariant } from 'helpers/contributions';
 import type { ExistingPaymentMethod } from 'helpers/forms/existingPaymentMethods/existingPaymentMethods';
 import { getSettings } from 'helpers/globalsAndSwitches/globals';
 import type { Settings } from 'helpers/globalsAndSwitches/settings';
+import { Country, CountryGroup } from 'helpers/internationalisation';
 import type { IsoCountry } from 'helpers/internationalisation/country';
 import type { CountryGroupId } from 'helpers/internationalisation/countryGroup';
 import type { IsoCurrency } from 'helpers/internationalisation/currency';
 import { detect as detectCurrency } from 'helpers/internationalisation/currency';
-import { Country, CountryGroup } from 'helpers/internationalisation/helpers';
 import type {
 	Campaign,
 	ReferrerAcquisitionData,
