+++ support-frontend/assets/pages/promotion-terms/promotionTerms.tsx
@@ -1,7 +1,7 @@
 import Footer from 'components/footerCompliant/Footer';
 import Header from 'components/headers/header/header';
 import Page from 'components/page/page';
-import { CountryGroup } from 'helpers/internationalisation/helpers';
+import { CountryGroup } from 'helpers/internationalisation';
 import { setUpTrackingAndConsents } from 'helpers/page/page';
 import type { PromotionTerms } from 'helpers/productPrice/promotions';
 import {
