+++ support-frontend/assets/pages/promotion-terms/weeklyTerms.tsx
@@ -1,8 +1,8 @@
 import { SansParagraph } from 'components/text/text';
+import { CountryGroup } from 'helpers/internationalisation';
 import type { CountryGroupName } from 'helpers/internationalisation/countryGroup';
 import { International } from 'helpers/internationalisation/countryGroup';
 import { extendedGlyph } from 'helpers/internationalisation/currency';
-import { CountryGroup } from 'helpers/internationalisation/helpers';
 import {
 	Annual,
 	postIntroductorySixForSixBillingPeriod,
